+++ backend/src/main/java/com/ssafy/api/service/FriendshipRepository.java
@@ -11,13 +11,13 @@ import com.ssafy.api.model.Friendship;
 @Repository
 public interface FriendshipRepository extends JpaRepository<Friendship, Integer>{
 
-	Optional<Friendship> findByFriend1IDAndFriend2ID(int user1id, int user2id);
+	Optional<Friendship> findByFriend1IDAndFriend2ID(long user1id, long user2id);
 
-	List<Friendship> findByFriend1ID(int userID);
+	List<Friendship> findByFriend1ID(long userID);
 
-	List<Friendship> findByFriend2ID(int userID);
+	List<Friendship> findByFriend2ID(long userID);
 
-	void deleteByFriend1ID(int userID);
+	void deleteByFriend1ID(long userID);
 
-	void deleteByFriend2ID(int userID);
+	void deleteByFriend2ID(long userID);
 }
